pyproject.toml
[project]
name = "Robot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cv2-enumerate-cameras>=1.3.0",
    "opencv-python>=4.12.0.88",
    "paho-mqtt>=2.1.0",
    "pyside6>=6.9.2",
]
